rename web to plain
