+++ cli/src/sourcemaps/mod.rs
@@ -2,5 +2,5 @@ pub mod constant;
 pub mod content;
 pub mod hermes;
 pub mod inject;
+pub mod plain;
 pub mod source_pairs;
-pub mod web;
